🎨 Clean the root module and add some basic documentation.
