+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod cbor;
+pub mod cli;
 pub mod errors;
 pub mod interplanetary;
 pub mod local;
